+++ id_ca.h
@@ -11,7 +11,7 @@
 
 typedef struct
 {
-        long            planestart[3];
+        int32_t            planestart[3];
         word planelength[3];
         word width,height;
         char            name[16];
@@ -39,18 +39,18 @@ extern  char            *titleptr[8];
 extern  char            extension[5];
 
 extern byte grstarts[(NUMCHUNKS+1)*3];  // array of offsets in egagraph, -1 for sparse
-extern long      *audiostarts;  // array of offsets in audio / audiot
+extern int32_t      *audiostarts;  // array of offsets in audio / audiot
 
 //===========================================================================
 
 //#define CA_FarRead(handle,dest,length) (read(handle,dest,length)==length)
 //#define CA_FarWrite(handle,source,length) (write(handle,source,length)==length)
 boolean CA_LoadFile (const char *filename, memptr *ptr);
-boolean CA_WriteFile (const char *filename, void *ptr, long length);
+boolean CA_WriteFile (const char *filename, void *ptr, int32_t length);
 
-long CA_RLEWCompress (word *source, long length, word *dest, word rlewtag);
+int32_t CA_RLEWCompress (word *source, int32_t length, word *dest, word rlewtag);
 
-void CA_RLEWexpand (word *source, word *dest,long length, word rlewtag);
+void CA_RLEWexpand (word *source, word *dest, int32_t length, word rlewtag);
 
 void CA_Startup (void);
 void CA_Shutdown (void);
